pyproject.toml
[project]
name = "django-mailing"
version = "0.1.0"
description = ""
authors = [
    {name = "Artem Filippow",email = "a.filippov@qtech.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pypugjs (>=5.12.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
